remove gameDefault to class GameXO, fix PostMapping now hi can take JSON object from request
